- added some new config values - refactored BaseApplication: removed one-line methods - some minor changes

+++ include/Blobb.h
@@ -33,5 +33,5 @@ public:
 	}
 
 	void move(Ogre::Vector2 direction);
-	void jump(float force);
+	void jump(int force);
 };
